+++ blockchain/blockchain_sync/blockchain_sync.go
@@ -6,7 +6,7 @@ import (
 	"pandora-pay/gui"
 	"pandora-pay/helpers/generics"
 	"pandora-pay/helpers/multicast"
-	"pandora-pay/recovery"
+	"pandora-pay/helpers/recovery"
 	"time"
 )
