+++ blockchain/forging/forging_thread.go
@@ -11,8 +11,8 @@ import (
 	"pandora-pay/gui"
 	"pandora-pay/helpers/advanced_buffers"
 	"pandora-pay/helpers/generics"
+	"pandora-pay/helpers/recovery"
 	"pandora-pay/mempool"
-	"pandora-pay/recovery"
 	"strconv"
 	"sync/atomic"
 	"time"
