+++ mempool/mempool.go
@@ -12,8 +12,8 @@ import (
 	"pandora-pay/gui"
 	"pandora-pay/helpers"
 	"pandora-pay/helpers/generics"
+	"pandora-pay/helpers/recovery"
 	"pandora-pay/network/websocks/connection/advanced_connection_types"
-	"pandora-pay/recovery"
 	"pandora-pay/txs_validator"
 	"runtime"
 	"time"
