+++ network/server/node_tcp/tcp_server.default.go
@@ -15,12 +15,12 @@ import (
 	"pandora-pay/config"
 	"pandora-pay/config/globals"
 	"pandora-pay/gui"
+	"pandora-pay/helpers/recovery"
 	"pandora-pay/mempool"
 	"pandora-pay/network/banned_nodes"
 	"pandora-pay/network/connected_nodes"
 	"pandora-pay/network/known_nodes"
 	"pandora-pay/network/server/node_http"
-	"pandora-pay/recovery"
 	"pandora-pay/settings"
 	"pandora-pay/txs_builder"
 	"pandora-pay/txs_validator"
