+++ network/websocks/connection/advanced_connection.go
@@ -9,10 +9,10 @@ import (
 	"pandora-pay/config"
 	"pandora-pay/helpers"
 	"pandora-pay/helpers/generics"
+	"pandora-pay/helpers/recovery"
 	"pandora-pay/network/known_nodes/known_node"
 	"pandora-pay/network/websocks/connection/advanced_connection_types"
 	"pandora-pay/network/websocks/websock"
-	"pandora-pay/recovery"
 	"sync"
 	"sync/atomic"
 	"time"
