+++ network/websocks/websocket_server_notjs.go
@@ -6,10 +6,10 @@ package websocks
 import (
 	"net/http"
 	"pandora-pay/config"
+	"pandora-pay/helpers/recovery"
 	"pandora-pay/network/connected_nodes"
 	"pandora-pay/network/known_nodes"
 	"pandora-pay/network/websocks/websock"
-	"pandora-pay/recovery"
 	"sync/atomic"
 )
