+++ network/websocks/websockets.go
@@ -10,6 +10,7 @@ import (
 	"pandora-pay/config/globals"
 	"pandora-pay/gui"
 	"pandora-pay/helpers/multicast"
+	"pandora-pay/helpers/recovery"
 	"pandora-pay/mempool"
 	"pandora-pay/network/api/api_http"
 	"pandora-pay/network/api/api_websockets"
@@ -20,7 +21,6 @@ import (
 	"pandora-pay/network/websocks/connection"
 	"pandora-pay/network/websocks/connection/advanced_connection_types"
 	"pandora-pay/network/websocks/websock"
-	"pandora-pay/recovery"
 	"pandora-pay/settings"
 	"strconv"
 	"sync/atomic"
